Fix error when importing files without tags
When no tags were specified, no actual files were copied on import,
while a database entry was still created. This was due to the fact, that
the import file paths are derived from the specified tags. Now, files
without tags are simply imported into the root folder of the document
location.